pyproject.toml
[tool.poetry]
name = "speech-hypertuning"
version = "0.1.0"
description = ""
authors = ["Erik Ernst"]
license = "Apache-2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pre-commit = "^3.6.0"
ginpipe = "^0.0.1"
lightning = "^2.2.0.post0"
pandas = "^2.2.1"
soundfile = "^0.12.1"
s3prl = "^0.4.14"
torchinfo = "^1.8.0"
pytest = "^8.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
